+++ exomiser-core/src/main/java/org/monarchinitiative/exomiser/core/analysis/AnalysisPresetBuilder.java
@@ -63,17 +63,26 @@ class AnalysisPresetBuilder {
             FrequencySource.ESP_AFRICAN_AMERICAN, FrequencySource.ESP_ALL, FrequencySource.ESP_EUROPEAN_AMERICAN,
             FrequencySource.THOUSAND_GENOMES,
             FrequencySource.EXAC_AFRICAN_INC_AFRICAN_AMERICAN, FrequencySource.EXAC_AMERICAN,
-            FrequencySource.EXAC_EAST_ASIAN, FrequencySource.EXAC_FINNISH,
-            FrequencySource.EXAC_NON_FINNISH_EUROPEAN, FrequencySource.EXAC_SOUTH_ASIAN, FrequencySource.EXAC_OTHER,
+            FrequencySource.EXAC_EAST_ASIAN,
+            FrequencySource.EXAC_NON_FINNISH_EUROPEAN, FrequencySource.EXAC_SOUTH_ASIAN,
             FrequencySource.UK10K, FrequencySource.TOPMED,
             FrequencySource.GNOMAD_E_AFR, FrequencySource.GNOMAD_E_AMR,
-            //FrequencySource.GNOMAD_E_ASJ,
-            FrequencySource.GNOMAD_E_EAS, FrequencySource.GNOMAD_E_FIN,
-            FrequencySource.GNOMAD_E_NFE, FrequencySource.GNOMAD_E_OTH, FrequencySource.GNOMAD_E_SAS,
+            FrequencySource.GNOMAD_E_EAS,
+            FrequencySource.GNOMAD_E_NFE,
+            FrequencySource.GNOMAD_E_SAS,
             FrequencySource.GNOMAD_G_AFR, FrequencySource.GNOMAD_G_AMR,
+            FrequencySource.GNOMAD_G_EAS,
+            FrequencySource.GNOMAD_G_NFE,
+            FrequencySource.GNOMAD_G_SAS
+            // small/bottle-necked populations not suitable for allele frequency filtering
+            //FrequencySource.EXAC_FINNISH,
+            //FrequencySource.EXAC_OTHER,
+            //FrequencySource.GNOMAD_E_ASJ,
+            //FrequencySource.GNOMAD_E_FIN,
+            //FrequencySource.GNOMAD_E_OTH,
             //FrequencySource.GNOMAD_G_ASJ,
-            FrequencySource.GNOMAD_G_EAS, FrequencySource.GNOMAD_G_FIN,
-            FrequencySource.GNOMAD_G_NFE, FrequencySource.GNOMAD_G_OTH, FrequencySource.GNOMAD_G_SAS
+            //FrequencySource.GNOMAD_G_FIN,
+            //FrequencySource.GNOMAD_G_OTH,
     );
 
     private static final HiPhiveOptions HI_PHIVE_OPTIONS = HiPhiveOptions.builder()