+++ exomiser-core/src/test/java/org/monarchinitiative/exomiser/core/analysis/util/acmg/Acmg2020PointsBasedClassifierTest.java
@@ -1,23 +1,3 @@
-/*
- * The Exomiser - A tool to annotate and prioritize genomic variants
- *
- * Copyright (c) 2016-2021 Queen Mary University of London.
- * Copyright (c) 2012-2016 Charité Universitätsmedizin Berlin and Genome Research Ltd.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
 package org.monarchinitiative.exomiser.core.analysis.util.acmg;
 
 import org.junit.jupiter.api.Test;
@@ -27,9 +7,10 @@ import org.junit.jupiter.params.provider.CsvSource;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
-class Acgs2020ClassifierTest {
+class Acmg2020PointsBasedClassifierTest {
 
-    private final Acgs2020Classifier instance = new Acgs2020Classifier();
+
+    private final AcmgEvidenceClassifier instance = new Acmg2020PointsBasedClassifier();
 
     private AcmgEvidence parseAcmgEvidence(String criteria) {
         AcmgEvidence.Builder acmgEvidenceBuilder = AcmgEvidence.builder();
@@ -40,7 +21,6 @@ class Acgs2020ClassifierTest {
                 AcmgCriterion.Evidence evidence = AcmgCriterion.Evidence.parseValue(criteriaModifier[1]);
                 acmgEvidenceBuilder.add(acmgCriterion, evidence);
             } else {
-//                AcmgCriterion acmgCriterion = AcmgCriterion.valueOf(criterion);
                 acmgEvidenceBuilder.add(acmgCriterion);
             }
         }
@@ -89,12 +69,14 @@ class Acgs2020ClassifierTest {
 
     @ParameterizedTest
     @CsvSource({
+            "PVS1 PP1",
 //      (c)  >=2 Strong
             "PS1 PS2",
-            "PVS1 PP1", // updated rule https://clinicalgenome.org/site/assets/files/5182/pm2_-_svi_recommendation_-_approved_sept2020.pdf
+            "PS1 PS2 PP1",
 //      (b)  1 Strong (PS1–PS4) AND
 //             1–2 moderate (PM1–PM6) OR
             "PS1 PM1 PM2",
+            "PS1 PM1",
 //             ≥2 supporting (PP1–PP5)
             "PS1 PP1 PP2",
             "PS1 PP1 PP2 PP3",
@@ -107,7 +89,7 @@ class Acgs2020ClassifierTest {
 //           1 Moderate (PM1–PM6) AND ≥4 supporting (PP1–PP5)
             "PM1 PP1 PP2 PP3 PP4",
             "PM1 PP1 PP2 PP3 PP4 PP5",
-            "PM1 PP1 PP2 PP3", // VUS_Hot - we can only assign 3 out of
+            //PM2, PP3_Strong, PP4, PP5_Strong
     })
     void classifiesLikelyPathogenic(String criteria) {
         AcmgEvidence acmgEvidence = parseAcmgEvidence(criteria);
@@ -116,38 +98,10 @@ class Acgs2020ClassifierTest {
 
     @ParameterizedTest
     @CsvSource({
-//            (i)  1 Stand-alone (BA1)
-            "BA1",
-            "BA1 BS1 BP1",
-//            OR (ii)  ≥2 Strong (BS1–BS4)
-            "BS1 BS2",
-            "BS1 BS2 BS3"
-    })
-    void classifiesBenign(String criteria) {
-        AcmgEvidence acmgEvidence = parseAcmgEvidence(criteria);
-        assertThat(instance.classify(acmgEvidence), equalTo(AcmgClassification.BENIGN));
-    }
-
-    @ParameterizedTest
-    @CsvSource({
-//            (i)   1 Strong (BS1–BS4) and 1 supporting (BP1–BP7)
-            "BS1 BP1",
-//            OR (ii)  ≥2 Supporting (BP1–BP7)
-            "BP1 BP2",
-            "BP1 BP2 BP3",
-    })
-    void classifiesLikelyBenign(String criteria) {
-        AcmgEvidence acmgEvidence = parseAcmgEvidence(criteria);
-        assertThat(instance.classify(acmgEvidence), equalTo(AcmgClassification.LIKELY_BENIGN));
-    }
-
-    @ParameterizedTest
-    @CsvSource({
 //            (i)  Other criteria shown above are not met
-//            "PVS1 PP1", //
             "PS1 PP1", // VUS_Hot
             "PM1 PM2 PP1", // VUS_Hot
-//            "PM1 PP1 PP2 PP3", // VUS_Hot - classified as LP
+            "PM1 PP1 PP2 PP3", // VUS_Hot - classified as LP
             "PS1", // VUS_Warm
             "PM1 PM2", // VUS_Warm
             "PM1 PP1 PP2", // VUS_Warm
@@ -159,17 +113,78 @@ class Acgs2020ClassifierTest {
             "PP1", // VUS_Cold
 //            OR  (ii)   the criteria for benign and pathogenic are contradictory
             "PS1 BP1",
-            "BS1 BP1 PVS1 PS1",
-            "BA1 PVS1 PM1",
+//            "BS1 BP1 PVS1 PS1", // This is LP (12P, 5B = 7 points)
+//            "BA1 PVS1 PM1", // is this possible?
             "BP1 BP2 BP3 PM1 PP1 PP2 PP3 PP4",
+            "PP1 BP1",
     })
     void classifiesUncertain(String criteria) {
         AcmgEvidence acmgEvidence = parseAcmgEvidence(criteria);
         assertThat(instance.classify(acmgEvidence), equalTo(AcmgClassification.UNCERTAIN_SIGNIFICANCE));
     }
 
+    @ParameterizedTest
+    @CsvSource({
+//            (i)   1 Strong (BS1–BS4) and 1 supporting (BP1–BP7)
+            "BS1 BP1",
+//            OR (ii)  ≥2 Supporting (BP1–BP7)
+            "BP1 BP2",
+            "BP4",
+            "BP4_Moderate", // This is possible
+            "BP1 BP2 BP3",
+            "BP4_Moderate BP6",
+    })
+    void classifiesLikelyBenign(String criteria) {
+        AcmgEvidence acmgEvidence = parseAcmgEvidence(criteria);
+        assertThat(instance.classify(acmgEvidence), equalTo(AcmgClassification.LIKELY_BENIGN));
+    }
+
+    @ParameterizedTest
+    @CsvSource({
+//            (i)  1 Stand-alone (BA1)
+            "BA1",
+            "BA1 BS1 BP1",
+//            OR (ii)  ≥2 Strong (BS1–BS4)
+            "BS1 BS2",
+            "BS1 BS2 BS3",
+            "PS3 PS4 PP1_Supporting PM3 PP4 BA1" // would be a warm VUS (4 points) if not for the BA1 benign hard-filter
+    })
+    void classifiesBenign(String criteria) {
+        AcmgEvidence acmgEvidence = parseAcmgEvidence(criteria);
+        assertThat(instance.classify(acmgEvidence), equalTo(AcmgClassification.BENIGN));
+    }
+
     @Test
     void noEvidenceClassifiesUnknown() {
         assertThat(instance.classify(AcmgEvidence.empty()), equalTo(AcmgClassification.UNCERTAIN_SIGNIFICANCE));
     }
+
+
+    @ParameterizedTest
+    @CsvSource({
+            "11, PATHOGENIC",
+            "10, PATHOGENIC",
+            "9, LIKELY_PATHOGENIC",
+            "8, LIKELY_PATHOGENIC",
+            "7, LIKELY_PATHOGENIC",
+            "6, LIKELY_PATHOGENIC",
+            "5, UNCERTAIN_SIGNIFICANCE",
+            "4, UNCERTAIN_SIGNIFICANCE",
+            "3, UNCERTAIN_SIGNIFICANCE",
+            "2, UNCERTAIN_SIGNIFICANCE",
+            "1, UNCERTAIN_SIGNIFICANCE",
+            "0, UNCERTAIN_SIGNIFICANCE",
+            "-1, LIKELY_BENIGN",
+            "-2, LIKELY_BENIGN",
+            "-3, LIKELY_BENIGN",
+            "-4, LIKELY_BENIGN",
+            "-5, LIKELY_BENIGN",
+            "-6, LIKELY_BENIGN",
+            "-7, BENIGN",
+            "-8, BENIGN",
+    })
+    void testPointsClassification(int points, AcmgClassification acmgClassification) {
+        assertThat(new Acmg2020PointsBasedClassifier().classification(points), equalTo(acmgClassification));
+    }
+
 }