+++ exomiser-core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # The Exomiser - Core Library Changelog
 
+## 13.3.0 2023-10-17
+
+- Updated Jannovar version to 0.41 to fix incorrect MT codon table usage [#521](https://github.com/exomiser/Exomiser/issues/521)
+- Downgraded PM2 - PM2_Supporting for variants lacking frequency information [#502](https://github.com/exomiser/Exomiser/issues/502).
+- Updated Acgs2020Classifier and Acmg2015Classifier to allow for PVS1 and PM2_Supporting to be sufficient to trigger LIKELY_PATHOGENIC
+- Updated AcmgEvidence to fit a Bayesian points-based system [#514](https://github.com/exomiser/Exomiser/issues/514)
+- Removed ASJ, FIN, OTH ExAC and gnomAD populations from presets and examples [#513](https://github.com/exomiser/Exomiser/issues/513).
+- Fix for regression causing `<INV>` variants to be incorrectly down-ranked 
+- Fix for issue [#486](https://github.com/exomiser/Exomiser/issues/486) where VCF output includes whitespace in INFO field.
+- Logs will now display elapsed time correctly if an analysis runs over an hour (!).
+- Updated exomiser-phenotype-data to take annotations from phenotype.hpoa [#351](https://github.com/exomiser/Exomiser/issues/351), [#373](https://github.com/exomiser/Exomiser/issues/373), [#379](https://github.com/exomiser/Exomiser/issues/379)
+- Updated application.properties and ResourceConfigurationProperties to remove unused fields.
+- Updated DiseaseInheritanceCacheReader and DiseasePhenotypeReader to parse phenotype.hpoa file
+- Updated DiseaseResourceConfig to use hpoa resource
+
+New APIs:
+
+- New `Acmg2020PointsBasedClassifier` class
+- New `FrequencyData.size()` method
+- New `FrequencyData.isEmpty()` method
+- New `FrequencyData.containsFrequencySource()` method
+- New `FrequencyData.getMaxFreqForPopulation()` method
+- New `FrequencySource.NON_FOUNDER_POPS` method
+
 ## 13.2.1 2023-06-30
 
 - Fix for bug where all `<INS>` structural variants were given a maximal variant score of 1.0 regardless of their position on a transcript.